kernel: grants: use one usize counter
The core kernel manages upcalls and allowed buffers for grants. Before,
three usize values were used to store the number of upcalls and allowed
buffers per grant. Now, we use a single usize and use 8 bit fields
inside of the usize.

This also motivated expanding the KernelManagedLayout struct to handle
more of the low-level operation.